ALL: Finished ManagedSurface's blendBlitFrom
I made TransparentSurface's blit function for ManagedSurface under the
name of blendBlitFrom, and made the backend function, blendBlitUnfiltered
in graphics/blit-alpha.cpp (TransparentSurface also uses this function
now). The blit function in RenderedImage from the sword25 engine was also
changed, as it didn't work if a pPartRect was passed in.